requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
python-multipart
jinja2
